[build] fix LLVM compilation issues (#13058)
- Fix Clang -Wshorten-64-to-32 in mac_frame.cpp and cli_ping.cpp

+++ examples/platforms/utils/mac_frame.cpp
@@ -406,7 +406,8 @@ otError otMacFrameProcessTxSfd(otRadioFrame *aFrame, uint64_t aRadioTime, otRadi
 #if OPENTHREAD_CONFIG_MAC_CSL_RECEIVER_ENABLE
     if (static_cast<Mac::Frame *>(aFrame)->Has<Mac::CslIe>()) // CSL IE should be filled for every transmit attempt
     {
-        otMacFrameSetCslIe(aFrame, aRadioContext->mCslPeriod, ComputeCslPhase(aRadioTime, aRadioContext));
+        otMacFrameSetCslIe(aFrame, aRadioContext->mCslPeriod,
+                           ComputeCslPhase(static_cast<uint32_t>(aRadioTime), aRadioContext));
     }
 #endif
 #if OPENTHREAD_CONFIG_TIME_SYNC_ENABLE
